Apply very_good_analysis lint rules
