+++ lib/model/alarm_settings.dart
@@ -1,4 +1,45 @@
+import 'package:flutter/widgets.dart';
+
+/// [AlarmSettings] is a model that contains all the settings to customize
+/// and set an alarm.
+@immutable
 class AlarmSettings {
+  /// Model that contains all the settings to customize and set an alarm.
+  ///
+  ///
+  /// Note that if you want to show a notification when alarm is triggered, both
+  ///  [notificationTitle] and [notificationBody] must not be null nor empty.
+  const AlarmSettings({
+    required this.id,
+    required this.dateTime,
+    required this.assetAudioPath,
+    required this.notificationTitle,
+    required this.notificationBody,
+    this.loopAudio = true,
+    this.vibrate = true,
+    this.volume,
+    this.fadeDuration = 0.0,
+    this.enableNotificationOnKill = true,
+    this.androidFullScreenIntent = true,
+  });
+
+  /// Constructs an `AlarmSettings` instance from the given JSON data.
+  factory AlarmSettings.fromJson(Map<String, dynamic> json) => AlarmSettings(
+        id: json['id'] as int,
+        dateTime: DateTime.fromMicrosecondsSinceEpoch(json['dateTime'] as int),
+        assetAudioPath: json['assetAudioPath'] as String,
+        loopAudio: json['loopAudio'] as bool,
+        vibrate: json['vibrate'] as bool? ?? true,
+        volume: json['volume'] as double?,
+        fadeDuration: json['fadeDuration'] as double,
+        notificationTitle: json['notificationTitle'] as String? ?? '',
+        notificationBody: json['notificationBody'] as String? ?? '',
+        enableNotificationOnKill:
+            json['enableNotificationOnKill'] as bool? ?? true,
+        androidFullScreenIntent:
+            json['androidFullScreenIntent'] as bool? ?? true,
+      );
+
   /// Unique identifier assiocated with the alarm. Cannot be 0 or -1;
   final int id;
 
@@ -7,15 +48,24 @@ class AlarmSettings {
 
   /// Path to audio asset to be used as the alarm ringtone. Accepted formats:
   ///
-  /// * **Project asset**: Specifies an asset bundled with your Flutter project. Use this format for assets that are included in your project's `pubspec.yaml` file.
+  /// * **Project asset**: Specifies an asset bundled with your Flutter project.
+  ///  Use this format for assets that are included in your project's
+  /// `pubspec.yaml` file.
   ///  Example: `assets/audio.mp3`.
-  /// * **Absolute file path**: Specifies a direct file system path to the audio file. This format is used for audio files stored outside the Flutter project, such as files saved in the device's internal or external storage.
+  /// * **Absolute file path**: Specifies a direct file system path to the
+  /// audio file. This format is used for audio files stored outside the
+  /// Flutter project, such as files saved in the device's internal
+  /// or external storage.
   ///  Example: `/path/to/your/audio.mp3`.
-  /// * **Relative file path**: Specifies a file path relative to a predefined base directory in the app's internal storage. This format is convenient for referring to files that are stored within a specific directory of your app's internal storage without needing to specify the full path.
+  /// * **Relative file path**: Specifies a file path relative to a predefined
+  /// base directory in the app's internal storage. This format is convenient
+  /// for referring to files that are stored within a specific directory of
+  /// your app's internal storage without needing to specify the full path.
   ///  Example: `Audios/audio.mp3`.
   ///
-  /// If you want to use aboslute or relative file path, you must request android storage
-  /// permission and add the following permission to your `AndroidManifest.xml`:
+  /// If you want to use aboslute or relative file path, you must request
+  /// android storage permission and add the following permission to your
+  /// `AndroidManifest.xml`:
   /// `android.permission.READ_EXTERNAL_STORAGE`
   final String assetAudioPath;
 
@@ -30,11 +80,13 @@ class AlarmSettings {
 
   /// Specifies the system volume level to be set at the designated [dateTime].
   ///
-  /// Accepts a value between 0 (mute) and 1 (maximum volume). When the alarm is triggered at [dateTime],
-  /// the system volume adjusts to this specified level. Upon stopping the alarm, the system volume reverts
+  /// Accepts a value between 0 (mute) and 1 (maximum volume).
+  /// When the alarm is triggered at [dateTime], the system volume adjusts to
+  /// this specified level. Upon stopping the alarm, the system volume reverts
   /// to its prior setting.
   ///
-  /// If left unspecified or set to `null`, the current system volume at the time of the alarm will be used.
+  /// If left unspecified or set to `null`, the current system volume
+  /// at the time of the alarm will be used.
   /// Defaults to `null`.
   final double? volume;
 
@@ -57,7 +109,8 @@ class AlarmSettings {
   /// when android alarm notification is triggered. Enabled by default.
   final bool androidFullScreenIntent;
 
-  /// Returns a hash code for this `AlarmSettings` instance using Jenkins hash function.
+  /// Returns a hash code for this `AlarmSettings` instance using
+  /// Jenkins hash function.
   @override
   int get hashCode {
     var hash = 0;
@@ -76,42 +129,6 @@ class AlarmSettings {
 
     return hash;
   }
-
-  /// Model that contains all the settings to customize and set an alarm.
-  ///
-  ///
-  /// Note that if you want to show a notification when alarm is triggered,
-  /// both [notificationTitle] and [notificationBody] must not be null nor empty.
-  const AlarmSettings({
-    required this.id,
-    required this.dateTime,
-    required this.assetAudioPath,
-    this.loopAudio = true,
-    this.vibrate = true,
-    this.volume,
-    this.fadeDuration = 0.0,
-    required this.notificationTitle,
-    required this.notificationBody,
-    this.enableNotificationOnKill = true,
-    this.androidFullScreenIntent = true,
-  });
-
-  /// Constructs an `AlarmSettings` instance from the given JSON data.
-  factory AlarmSettings.fromJson(Map<String, dynamic> json) => AlarmSettings(
-        id: json['id'] as int,
-        dateTime: DateTime.fromMicrosecondsSinceEpoch(json['dateTime'] as int),
-        assetAudioPath: json['assetAudioPath'] as String,
-        loopAudio: json['loopAudio'] as bool,
-        vibrate: json['vibrate'] as bool? ?? true,
-        volume: json['volume'] as double?,
-        fadeDuration: json['fadeDuration'] as double,
-        notificationTitle: json['notificationTitle'] as String? ?? '',
-        notificationBody: json['notificationBody'] as String? ?? '',
-        enableNotificationOnKill:
-            json['enableNotificationOnKill'] as bool? ?? true,
-        androidFullScreenIntent:
-            json['androidFullScreenIntent'] as bool? ?? true,
-      );
 
   /// Creates a copy of `AlarmSettings` but with the given fields replaced with
   /// the new values.
@@ -163,10 +180,11 @@ class AlarmSettings {
   /// Returns all the properties of `AlarmSettings` for debug purposes.
   @override
   String toString() {
-    Map<String, dynamic> json = toJson();
-    json['dateTime'] = DateTime.fromMicrosecondsSinceEpoch(json['dateTime']);
+    final json = toJson();
+    json['dateTime'] =
+        DateTime.fromMicrosecondsSinceEpoch(json['dateTime'] as int);
 
-    return "AlarmSettings: ${json.toString()}";
+    return 'AlarmSettings: $json';
   }
 
   /// Compares two AlarmSettings.
