+++ lib/src/android_alarm.dart
@@ -1,22 +1,29 @@
 import 'dart:async';
 import 'package:alarm/alarm.dart';
-import 'package:alarm/service/storage.dart';
+import 'package:alarm/model/alarm_settings.dart';
+import 'package:alarm/service/alarm_storage.dart';
+import 'package:alarm/utils/alarm_exception.dart';
 import 'package:flutter/services.dart';
 
 /// Uses method channel to interact with the native platform.
 class AndroidAlarm {
+  /// Method channel for the alarm.
   static const platform = MethodChannel('com.gdelataillade.alarm/alarm');
 
+  /// Whether there are other alarms set.
   static bool get hasOtherAlarms => AlarmStorage.getSavedAlarms().length > 1;
 
+  /// Initializes the method channel.
   static Future<void> init() async {
     platform.setMethodCallHandler(handleMethodCall);
   }
 
+  /// Handles the method call from the native platform.
   static Future<dynamic> handleMethodCall(MethodCall call) async {
     try {
       if (call.method == 'alarmRinging') {
-        int id = call.arguments['id'];
+        final arguments = call.arguments as Map<String, dynamic>;
+        final id = arguments['id'] as int;
         final settings = Alarm.getAlarm(id);
         if (settings != null) Alarm.ringStream.add(settings);
       }
@@ -25,7 +32,7 @@ class AndroidAlarm {
     }
   }
 
-  /// Schedules a native alarm with given [alarmSettings] with its notification.
+  /// Schedules a native alarm with given [settings] with its notification.
   static Future<bool> set(
     AlarmSettings settings,
     void Function()? onRing,
@@ -74,15 +81,15 @@ class AndroidAlarm {
   /// Sends the message `stop` to the isolate so the audio player
   /// can stop playing and dispose.
   static Future<bool> stop(int id) async {
-    final res = await platform.invokeMethod('stopAlarm', {'id': id});
+    final res = await platform.invokeMethod('stopAlarm', {'id': id}) as bool;
     if (res) alarmPrint('Alarm with id $id stopped');
-    if (!hasOtherAlarms) stopNotificationOnKillService();
+    if (!hasOtherAlarms) await stopNotificationOnKillService();
     return res;
   }
 
   /// Checks if the alarm with given [id] is ringing.
   static Future<bool> isRinging(int id) async {
-    final res = await platform.invokeMethod('isRinging', {'id': id});
+    final res = await platform.invokeMethod('isRinging', {'id': id}) as bool;
     return res;
   }
