+++ lib/utils/alarm_exception.dart
@@ -1,0 +1,11 @@
+/// Custom exception for the alarm.
+class AlarmException implements Exception {
+  /// Creates an [AlarmException] with the given error [message].
+  const AlarmException(this.message);
+
+  /// Exception message.
+  final String message;
+
+  @override
+  String toString() => message;
+}
