+++ pubspec.yaml
@@ -16,7 +16,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.1
+  very_good_analysis: ^5.1.0
 
 flutter:
   assets:
